requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2018.6
requests==2.20.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24
